requirements.txt
pdfplumber==0.11.4
PyMuPDF==1.24.9
python-docx==1.1.2
Pillow==10.4.0
pytesseract==0.3.13
